csr_jss/csr/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
backports.csv==1.0.7
beautifulsoup4==4.9.3
biopython==1.78
blis==0.7.4
boto3==1.17.43
botocore==1.20.43
catalogue==2.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
colorama==0.4.4
coloredlogs==15.0
cryptography==3.4.7
cssselect==1.1.0
cymem==2.0.5
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-import-export==2.5.0
django-js-asset==1.2.2
docopt==0.6.2
docxtpl==0.11.4
et-xmlfile==1.0.1
eutils==0.6.0
feedparser==6.0.2
filelock==3.0.12
future==0.18.2
habanero==0.7.4
htmldocx==0.0.4
humanfriendly==9.1
idna==2.10
iniconfig==1.1.1
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
metapub==0.5.5
more-itertools==8.7.0
murmurhash==1.0.5
mysqlclient==2.0.3
nltk==3.5
numpy==1.19.3
odfpy==1.4.1
openpyxl==3.0.5
ordered-set==4.0.2
packaging==20.9
pandas==1.1.4
pathy==0.4.0
Pattern==3.6
pdfminer.six==20201018
Pillow==8.2.0
pluggy==0.13.1
portend==2.7.1
preshed==3.0.5
psycopg2-binary==2.8.6
pubmed-lookup==0.2.3
py==1.10.0
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pyreadline==2.1
pytest==6.2.3
python-dateutil==2.8.1
python-docx==0.8.10
python-Levenshtein==0.12.2
pytz==2020.4
pywin32==300
PyYAML==5.3.1
regex==2021.3.17
requests==2.25.0
s3transfer==0.3.3
scipy==1.6.2
sgmllib3k==1.0.0
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.2
sqlparse==0.4.1
srsly==2.4.0
tablib==2.0.0
tabulate==0.8.9
tempora==4.0.1
thinc==8.0.2
toml==0.10.2
tox==3.23.0
tqdm==4.59.0
typer==0.3.2
Unidecode==1.2.0
urllib3==1.26.2
virtualenv==20.4.4
wasabi==0.8.2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zc.lockfile==2.0
